tests/wpt/web-platform-tests/html/canvas/offscreen/the-offscreen-canvas/2d.canvas.reference.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.canvas.reference</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.canvas.reference</h1>
<p class="desc">canvas refers back to its canvas</p>


<script>
var t = async_test("canvas refers back to its canvas");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

_assertSame(ctx.canvas, offscreenCanvas, "ctx.canvas", "offscreenCanvas");
t.done();

});
</script>
